+++ src/audio_super_resolution/__init__.py
@@ -2,6 +2,8 @@
 
 from .audiosr_backend import AUDIOSR_MODEL_NAMES, AUDIOSR_SAMPLE_RATE, AudiosrBackend
 from .config import InferenceConfig, default_model_cache_dir
+from .manifest import build_manifest, write_manifest
+from .models import ModelInfo, list_models
 from .quality import AudioQualityReport, format_quality_report, inspect_audio_quality
 from .resolver import (
     DEFAULT_AUDIO_EXTENSIONS,
@@ -22,17 +24,21 @@ __all__ = [
     "BackendInfo",
     "EnhancementResult",
     "InferenceConfig",
+    "ModelInfo",
     "PlannedEnhancement",
     "AudioQualityReport",
     "AUDIOSR_MODEL_NAMES",
     "AUDIOSR_SAMPLE_RATE",
     "available_backends",
+    "build_manifest",
     "discover_audio_files",
     "default_model_cache_dir",
     "format_quality_report",
     "get_backend",
     "inspect_audio_quality",
+    "list_models",
     "plan_enhancements",
+    "write_manifest",
 ]
 
 __version__ = "0.1.0"